refactor(server): move lambda sub-services/utils/constants/types into subfolders with conventional suffixes
Align the lambda driver split with the repo conventions:
- services/  + .service.ts  (LambdaAwsClientService, LambdaToolFunctionsService,
                              LambdaLayerManagerService, LambdaExecutorManagerService)
- utils/     + .util.ts     (compute-hashed-lambda-resource-name,
                              get-lambda-deps-layer-name, get-lambda-sdk-layer-name,
                              parse-lambda-log-result, reprefix-lambda-zip-entries)
- utils/__tests__/ + .util.spec.ts
- constants/ + .constant.ts (lambda-driver.constant.ts)
- types/     + .type.ts     (lambda-driver.type.ts)

Class names get the matching Service suffix and constructor types are
updated accordingly. No behavioral change.
