+++ packages/twenty-server/src/engine/core-modules/logic-function/logic-function-drivers/drivers/lambda/utils/__tests__/compute-hashed-lambda-resource-name.util.spec.ts
@@ -1,4 +1,4 @@
-import { computeHashedLambdaResourceName } from 'src/engine/core-modules/logic-function/logic-function-drivers/drivers/lambda/utils/compute-hashed-lambda-resource-name';
+import { computeHashedLambdaResourceName } from 'src/engine/core-modules/logic-function/logic-function-drivers/drivers/lambda/utils/compute-hashed-lambda-resource-name.util';
 
 describe('computeHashedLambdaResourceName', () => {
   it('returns prefix concatenated with a 12-character sha256 hex checksum of the contents', () => {
